requirements.txt

fastapi
uvicorn
python-dotenv
google-generativeai>=0.3.0
